LPS-178732 feature-flag-web: uses @Clusterable annotation
