+++ modules/apps/feature-flag/feature-flag-web/src/main/java/com/liferay/feature/flag/web/internal/manager/FeatureFlagPreferencesManager.java
@@ -14,155 +14,18 @@
 
 package com.liferay.feature.flag.web.internal.manager;
 
-import com.liferay.feature.flag.web.internal.constants.FeatureFlagConstants;
-import com.liferay.portal.kernel.cluster.ClusterExecutor;
-import com.liferay.portal.kernel.cluster.ClusterRequest;
-import com.liferay.portal.kernel.module.framework.service.IdentifiableOSGiService;
-import com.liferay.portal.kernel.module.framework.service.IdentifiableOSGiServiceUtil;
-import com.liferay.portal.kernel.portlet.PortalPreferences;
-import com.liferay.portal.kernel.service.PortalPreferencesLocalService;
-import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.MethodHandler;
-import com.liferay.portal.kernel.util.MethodKey;
-import com.liferay.portal.kernel.util.PortletKeys;
-import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portlet.PortalPreferencesWrapper;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.BiConsumer;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Drew Brokke
  */
-@Component(
-	service = {
-		FeatureFlagPreferencesManager.class, IdentifiableOSGiService.class
-	}
-)
-public class FeatureFlagPreferencesManager implements IdentifiableOSGiService {
+public interface FeatureFlagPreferencesManager {
 
 	public void addSubscriber(
-		long companyId, BiConsumer<String, Boolean> biConsumer) {
+		long companyId, BiConsumer<String, Boolean> biConsumer);
 
-		_subscribersMap.compute(
-			companyId,
-			(key, value) -> {
-				if (value == null) {
-					value = new ArrayList<>();
-				}
+	public Boolean isEnabled(long companyId, String key);
 
-				value.add(biConsumer);
-
-				return value;
-			});
-	}
-
-	@Override
-	public String getOSGiServiceIdentifier() {
-		return FeatureFlagPreferencesManager.class.getName();
-	}
-
-	public Boolean isEnabled(long companyId, String key) {
-		if (Validator.isNull(
-				_portalPreferencesLocalService.fetchPortalPreferences(
-					companyId, PortletKeys.PREFS_OWNER_TYPE_COMPANY))) {
-
-			return null;
-		}
-
-		PortalPreferences portalPreferences = _getPortalPreferences(companyId);
-
-		String value = portalPreferences.getValue(_NAMESPACE, key);
-
-		if (value == null) {
-			return null;
-		}
-
-		return GetterUtil.getBoolean(value);
-	}
-
-	public void setEnabled(long companyId, String key, boolean enabled) {
-		PortalPreferences portalPreferences = _getPortalPreferences(companyId);
-
-		portalPreferences.setValue(_NAMESPACE, key, String.valueOf(enabled));
-
-		_portalPreferencesLocalService.updatePreferences(
-			companyId, PortletKeys.PREFS_OWNER_TYPE_COMPANY, portalPreferences);
-
-		_notifySubscribers(companyId, key, enabled);
-
-		_notifyCluster(companyId, key, enabled);
-	}
-
-	private static void _onNotify(
-		String osgiServiceIdentifier, long companyId, String key,
-		boolean enabled) {
-
-		FeatureFlagPreferencesManager ploEntryModelListener =
-			(FeatureFlagPreferencesManager)
-				IdentifiableOSGiServiceUtil.getIdentifiableOSGiService(
-					osgiServiceIdentifier);
-
-		ploEntryModelListener._notifySubscribers(companyId, key, enabled);
-	}
-
-	private PortalPreferences _getPortalPreferences(long companyId) {
-		PortalPreferencesWrapper portalPreferencesWrapper =
-			(PortalPreferencesWrapper)
-				_portalPreferencesLocalService.getPreferences(
-					companyId, PortletKeys.PREFS_OWNER_TYPE_COMPANY);
-
-		return portalPreferencesWrapper.getPortalPreferencesImpl();
-	}
-
-	private void _notifyCluster(long companyId, String key, boolean enabled) {
-		if (!_clusterExecutor.isEnabled()) {
-			return;
-		}
-
-		MethodHandler methodHandler = new MethodHandler(
-			_onNotifyMethodKey, getOSGiServiceIdentifier(), companyId, key,
-			enabled);
-
-		ClusterRequest clusterRequest = ClusterRequest.createMulticastRequest(
-			methodHandler, true);
-
-		clusterRequest.setFireAndForget(true);
-
-		_clusterExecutor.execute(clusterRequest);
-	}
-
-	private void _notifySubscribers(
-		long companyId, String key, boolean enabled) {
-
-		List<BiConsumer<String, Boolean>> biConsumers =
-			_subscribersMap.getOrDefault(companyId, Collections.emptyList());
-
-		for (BiConsumer<String, Boolean> biConsumer : biConsumers) {
-			biConsumer.accept(key, enabled);
-		}
-	}
-
-	private static final String _NAMESPACE = FeatureFlagConstants.FEATURE_FLAG;
-
-	private static final MethodKey _onNotifyMethodKey = new MethodKey(
-		FeatureFlagPreferencesManager.class, "_onNotify", String.class,
-		Long.class, String.class, Boolean.class);
-
-	@Reference
-	private ClusterExecutor _clusterExecutor;
-
-	@Reference
-	private PortalPreferencesLocalService _portalPreferencesLocalService;
-
-	private final Map<Long, List<BiConsumer<String, Boolean>>> _subscribersMap =
-		new ConcurrentHashMap<>();
+	public void setEnabled(long companyId, String key, boolean enabled);
 
 }
